[PM-24279] clean up
