+++ src/Core/AdminConsole/OrganizationFeatures/Policies/Models/EmptyMetadataModel.cs
@@ -1,0 +1,6 @@
+﻿
+namespace Bit.Core.AdminConsole.OrganizationFeatures.Policies.Models;
+
+public record EmptyMetadataModel : IPolicyMetadataModel
+{
+}
