+++ src/Core/AdminConsole/OrganizationFeatures/Policies/Models/IPolicyMetadataModel.cs
@@ -1,0 +1,6 @@
+﻿
+namespace Bit.Core.AdminConsole.OrganizationFeatures.Policies.Models;
+
+public interface IPolicyMetadataModel
+{
+}
