+++ src/Core/AdminConsole/OrganizationFeatures/Policies/Models/OrganizationModelOwnershipPolicyModel.cs
@@ -1,0 +1,16 @@
+﻿
+namespace Bit.Core.AdminConsole.OrganizationFeatures.Policies.Models;
+
+public class OrganizationModelOwnershipPolicyModel : IPolicyMetadataModel
+{
+    public OrganizationModelOwnershipPolicyModel()
+    {
+    }
+
+    public OrganizationModelOwnershipPolicyModel(string? defaultUserCollectionName)
+    {
+        DefaultUserCollectionName = defaultUserCollectionName;
+    }
+
+    public string? DefaultUserCollectionName { get; set; }
+}
