+++ src/Core/AdminConsole/OrganizationFeatures/Policies/Models/SavePolicyModel.cs
@@ -1,0 +1,14 @@
+﻿
+
+using Bit.Core.AdminConsole.Models.Data;
+
+namespace Bit.Core.AdminConsole.OrganizationFeatures.Policies.Models;
+
+// Jimmy TODO: consider using VNext in these models
+public record SavePolicyModel(PolicyUpdate Data, IActingUser? PerformedBy, IPolicyMetadataModel Metadata)
+{
+    public PolicyUpdate Data { get; init; } = Data;
+    public IPolicyMetadataModel Metadata { get; init; } = Metadata;
+
+    public IActingUser? PerformedBy { get; init; } = PerformedBy;
+}
